Add Python 3 support - Switch from Attest to nose for tests since Attest does not support Python 3 - Unsure yet if this is tested properly and wether unicode strings are properly handled, but all tests pass - Fixed test regarding api key changing. (The user should not be able to login with their cookie if the secret key changes)
